+++ programs/greeting/src/entrypoint.rs
