+++ programs/greeting/src/state.rs
@@ -1,0 +1,8 @@
+use borsh::{BorshDeserialize, BorshSerialize};
+
+/// Define the type of state stored in accounts
+#[derive(BorshSerialize, BorshDeserialize, Debug)]
+pub struct GreetingAccount {
+    /// number of greetings
+    pub counter: u32,
+}
