+++ programs/scheduler/src/error.rs
