+++ programs/scheduler/src/lib.rs
